Hack: Workaround
Django executes a daemon thread for the wsgi server when you execute the
LiveServerTestCase (used by Behave). Unfortunately in our case, the
thread is not killed and I don't know why. But this workaround will
override the _teadDownClassInternal() method.

Issue, with this workaround, if we execute 100 scenarii, we will have
100 threads :-(

I will continue to investigate, but the problem seems to be in the code
of Behave or of the functional tests, because we don't have this issue
with the functional tests and unittests.